Properly separate atoms and molecules
